requirements.txt
flask==2.2.3
flask-cors==3.1.1
tensorflow==2.11.0
scikit-learn==1.0.2
numpy==1.23.4
pandas==1.5.1
Pillow==9.2.0
